Hashing of the message based on previous hash.
