+++ syslog/LogDestination.py
@@ -1,0 +1,32 @@
+class LogDestination(object):
+
+    def open(self):
+        """Open a connection to the target service
+
+        Should return False if opening fails"""
+        return True
+
+    def close(self):
+        """Close the connection to the target service"""
+        pass
+
+    def is_opened(self):
+        """Check if the connection to the target is able to receive messages"""
+        return True
+
+    def init(self, options):
+        """This method is called at initialization time
+
+        Should return false if initialization fails"""
+        return True
+
+    def deinit(self):
+        """This method is called at deinitialization time"""
+        pass
+
+    def send(self, msg):
+        """Send a message to the target service
+
+        It should return True to indicate success, False will suspend the
+        destination for a period specified by the time-reopen() option."""
+        return True
