--- syslog/blockchain.py
@@ -1,3 +1,0 @@
-# This file is supposed to contain the APIs that will interact 
-# with the blockchain
-
